third_party/blink/web_tests/external/wpt/soft-navigation-heuristics/detection/tentative/racing-soft-navigations.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>Two soft navigations racing each other.</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/resources/testdriver.js"></script>
    <script src="/resources/testdriver-vendor.js"></script>
    <script></script>
  </head>
  <body>
    <div id="slow-soft-navigation">Click here!</div>
    <div id="fast-soft-navigation">Click here!</div>

    <script>
      // This soft navigation is slow - it will wait 1s before it changes the URL.
      function slowSoftNavigation(t) {
        const greeting = document.createElement("div");
        greeting.textContent = "Hello, World.";
        document.body.appendChild(greeting);
        t.step_timeout(() => history.pushState({}, "", "/slow-soft-navigation"), 1000);
      }

      // This soft navigation is fast - it will change the URL immediately.
      function fastSoftNavigation() {
        const greeting = document.createElement("div");
        greeting.textContent = "Hello, World.";
        document.body.appendChild(greeting);
        history.pushState({}, "", "/fast-soft-navigation");
      }

      promise_test(async (t) => {
        // TODO(crbug.com/418839741) - Update this test if/when we also detect the slow navigation.

        document.getElementById("slow-soft-navigation").addEventListener("click", () => {
          slowSoftNavigation(t);
        });
        document
          .getElementById("fast-soft-navigation")
          .addEventListener("click", fastSoftNavigation);

        // Wait for the fast soft navigation to complete.
        const promise = new Promise((resolve) => {
          let entries = [];
          new PerformanceObserver((list, observer) => {
            entries.push(...list.getEntries());
            if (entries.length >= 1) {
              observer.disconnect();
              resolve(entries);
            }
          }).observe({ type: "soft-navigation" });
        });
        // Start both soft navigations in rapid succession.
        if (test_driver) {
          test_driver.click(document.getElementById("slow-soft-navigation"));
          test_driver.click(document.getElementById("fast-soft-navigation"));
        }

        // Notice that only the fast soft navigation is detected.
        const entries = await promise;
        assert_equals(entries.length, 1, "Expected just one soft navigation entry.");
        assert_equals(
          entries[0].name.replace(/.*\//, ""),
          "fast-soft-navigation",
          "First entry should be the fast soft navigation.",
        );

        // Wait for the effect of the slow soft navigation (URL change).
        // We're not detecting this soft navigation for now. See crbug.com/418839741.
        await t.step_wait(
          () => location.href.replace(/.*\//, "") === "slow-soft-navigation",
          "Expecting the slow navigation to eventually change the URL.",
        );
      }, "Two soft navigations that race each other should be detected correctly.");
    </script>
  </body>
</html>
